Revert "Merged in iulian07/handle32bit (pull request #56)"
This reverts commit 1e56ed928bf6b96ce7d60975eb103c7383e1aa12, reversing
changes made to e56f1210753d2d3f06e116ce5fae60c027c59d93.

+++ ReadHDF5.cpp
@@ -3539,10 +3539,7 @@ ErrorCode ReadHDF5::store_file_ids( Tag tag )
 
   CHECK_OPEN_HANDLES;
 
-  //typedef int tag_type;
-  typedef long tag_type;
-  // change it to be able to read much bigger files (long is 64 bits ...)
-
+  typedef int tag_type;
   tag_type* buffer = reinterpret_cast<tag_type*>(dataBuffer);
   const long buffer_size = bufferSize / sizeof(tag_type);
   for (IDMap::iterator i = idMap.begin(); i != idMap.end(); ++i) {
